services/requirements.txt
faiss-cpu
sentence-transformers
fastapi
uvicorn
numpy
requests
langchain
openai